Bug #16874873 "ROW SIZE TOO LARGE" ERROR THROWN 	      EVEN WHEN IT'S NOT PROBLEM -------
User is creating a table with many TEXT/BLOB fields 
and getting error which says that the row size 
is too large.

FIX
---
This behavior is expected because inoodb design
requires the record to fit in the B-tree leaf 
page and if unsuccessful we return error.During
the creation of BLOB or TEXT field we allocate
41 Bytes (dynamic or compressed row format) for
the field which is considered for row size 
calculation .During insertion if the BLOB or
TEXT is less than 40 Bytes  we store it internally 
with in these 41 bytes and if it is more than 40 
bytes,we store it off page and store 20 Bytes 
reference to the page.We have to update the 
documentation with this information . To give 
more information to the user we are adding extra 
information in the error log about the maximum
record size allowed,the present record size and
till what field we can accommodate so that the
record fits in the page. 

Approved by Kevin [#rb5384]