Bug#17748273: Failure in get_sort_by_table
The problem occurs when Item_field::fix_fields() resolves a column
reference from a list (find_item_in_list), and the field is an outer
reference. In this case, the depended_from field was not set, and the
used_tables property of the field was later calculated wrongly.

The solution is to set the depended_from field to the same value as
the item found in the list, by using the function mark_as_dependent().