Bug#18694751 72512: Non-aggregated query with set function in ORDER BY should be rejected
A non-aggregated query with an ORDER BY applied to it cannot contain
aggregate functions. This patch prevents such queries.

A new error message is added: ER_AGGREGATE_ORDER_NON_AGG_QUERY.

Notice that one test case with HAVING and an aggregated argument fails but
should not. This is because of a wrong resolve order in
SELECT_LEX::prepare().