This is a patch for bug#18107853 - XA LIST GETS CORRUPT, CRASH AND/OR HANG AND/OR ASSERTION
From user point of view an XA command that got XID value with too long
format ID part could cause crash or hang in the presence of concurrent
running XA commands.

The original test case attached to the bug report demonstrates problem
on 64-bit Windows. It assigns too long hex value to a format ID part of
XA xid, which leads to its overflow. This overflow condition wasn't
handled correctly within a parser that resulted in assignment of the
value -1 to the data member XID::formatID.

The reason for getting -1 from too long hex string value is that
my_strtoll() function, which is used to convert value from hex format,
returns the special value LLONG_MAX (0x7fffffffffffffff) in this case.
This 64-bit value later is casted to 32-bit unsigned long that leads to
discarding highest significant bits and in result the formatID gets value
0xffffffff (i.e. -1). Similar effect can be achieved on all platforms by
trying to assign decimal 18446744073709551615 to a format ID part of XA xid.

The xid value with format ID equal to -1 is interpreted by the implementation
of XA as a xid whose value is NULL. Further usage of such XID value leads to
error condition during operations on transaction_cache. This error condition
is expressed in the fact that the NULL xid value stored in the
transaction_cache can be modified later during executing of DML statement
within the same transaction. That is, if an user executes
              XA START 'xid1', 1, 0xfffffffffffffffffff;
              INSERT INTO some_table VALUES (some_value);
then during execution of INSERT statement the method XID_STATE::set_query_id
is called when an InnoDB transaction is registered  with the MySQL 2PC
coordinator and since the value of XID::formatID has value -1 as result of
overflow then existing xid value is considered as NULL xid and new xid value
is assigned. As result xid value which is already stored in the
transaction_cache is modified without recalculation of its hash key.
Modification of a value stored in the cache without a key recalculation could
lead to a crash or hang in case when several threads running XA-related
transaction are executed in parallel and all of them uses too long value that
overflow the values of formatID. Since it is difficult to create a reproducible
test that can lead to the crash this patch only provides test case that
prohibits execution of XA-transaction with value of formatID that resulted in
formatID equal to -1 in the past.

Full-blown checking of xid's formatID value requires significant changes
to source code and therefore has to be done in the scope of separate
bug/worklog. This patch takes simplistic approach to the problem and ensures
that formatID value passed from parser never gets negative value.