Fix for Bug #18770217 	DIFFERENT RESULT WITH SP ON 2ND EXECUTION OF QUERY WITH STRAIGHT_JOIN IN NOT IN We have this subquery predicate: NOT IN ( SELECT sq1_alias1.col_varchar_nokey FROM t3 AS sq1_alias1 JOIN ( t1 AS sq1_alias2 STRAIGHT_JOIN view_d AS sq1_alias3 ) WHERE ( SELECT col_int_key FROM t1 AS c_sq1_alias1 WHERE c_sq1_alias1.col_varchar_key = sq1_alias3.col_varchar_nokey ) ) which is handled with subquery materialization. It contains a WHERE clause involving a scalar subquery which returns two rows. Thus, it is expected that when we fill the temporary table, we hit and report an error. However, in the second execution of a stored procedure it is not the case. A necessary ingredient is the view column of view_d, used as outer reference in the scalar subquery. First execution: - view_d is merged - the view's column is resolved by fix_outer_field():           prev_subselect_item->used_tables_cache|=             (*reference)->used_tables();     // 1           prev_subselect_item->const_item_cache&=             (*reference)->const_item();           mark_as_dependent(thd, last_checked_context->select_lex, // 2
*reference is Item_direct_view_ref. Its depended_from is NULL at
(1). So its used_tables() returns the map of t2.col_varchar_nokey
(==4). This goes into used_tables_cache of the scalar subquery item,
which is correct - from the POV of the materialized subquery which
owns the scalar subquery, this scalar subquery depends on view_d, the
table of map 4 in the materialized subquery.
At (2), depended_from becomes "select#2" i.e. the materialized
subquery. It is correct: the item_direct_view_ref has the scalar
subquery as context, so it is an outer reference.

Second execution:
- in find_field_in_tables() we go through code specific of any not-first
execution:
  if (item->cached_table) <<<< this is true on non-first execution
  {
  ...
        /*
          If the field was an outer referencee, mark all selects using this
          sub query as dependent on the outer query
        */
        if (current_sel != last_select)
          mark_select_range_as_dependent(thd, last_select, current_sel,
                                         found, *ref, item);
This sets depended_from of the item_direct_view_ref to "select#2"
(correct).
- then we go into the same fix_outer_field() as in the first
execution, but note that this time we enter it with
depended_from!=NULL, so:
          prev_subselect_item->used_tables_cache|=
            (*reference)->used_tables();     // 1
adds OUTER_REF_TABLE_BIT to used_tables_cache of the scalar
subquery. That is wrong.
- then, this scalar subquery gets lost by make_join_select() because
this function (as it is operating for a materialized subquery)
excludes conditions using OUTER_REF_TABLE_BIT (see
join->allow_outer_refs).
- the condition being lost, it won't throw the expected "more than one
row" error.

Fix: use resolved_used_tables(). Indeed, prev_subselect_item is owned
by the qualifying query of the item. A second similar line is changed too.