Bug#18718491: MSI PACKING BROKEN WITH VS2013
This patch fixes problems building MSI packages with Visual Studio 2013.
VS2013 requires WiX Toolkit 3.8 and before this patch the code only worked
with 3.5 and 3.6.

The patch raises the minimum required version of CMake to 2.8.12 for Windows
only, as this version of CMake is required for VS2013. This also allows removing
a workaround for an old CMake bug (CPACK_MONOLITHIC_INSTALL).

The recommended way of building ZIP and MSI packages is:
1: Run CMake as normal, specifying VS2013 as generator.
2: cmake --build . --config (e.g debug) --target package
3: cmake --build . --config (e.g. debug) --target msi