Bug #18057562: PROXY USERS LOCKED OUT WHEN UNDERLYING PROXIED USER PASSWORD IS EXPIRED
Description: The proxy user can't execute any commands when
the proxied user's password is expired.

Analysis: During the authentication of a proxy user, the
"password_expired" variable is set to "password_expired"
value of the proxied user. Because of this, the proxy user
also can't execute any operations when the proxied user's
password is expired.

Fix: Allowing proxy users to execute operations even if
proxied user's password is expired. At the same time,
expired passwords should prevent the proxied user from
performing operations until SET PASSWORD is issued (this
behavior is already present).
As a fix, We are copying the acl_user password expired flag
to the security context. Before the acl_user gets updated.
So, that the current user password_expired flag will be
maintained.
We are checking for mpvio.acl_use->password_expired  
status in stead of acl_user, by this the older proxy 
clients(5.5 client) can connect and execute the quires 
successfully even the proxied user password expires.