WL7639 Ndb Active Active Manageability improvements
Implement status variables exposing last commit epochs for the server
and each session.

New status variables Ndb_last_commit_epoch_server and 
Ndb_last_commit_epoch_session are added to the MySQL server, allowing
SQL layer code to determine e.g.
  - What was the epoch of the last transaction committed by this server
    (Same info as embedded in SHOW ENGINE NDB STATUS output)
  - What was the epoch of the last transaction committed by this session

The session variant particularly is useful when determining when various
post-commit activities have completed on a transaction.  e.g. it could be 
used to determine k-safety round a replication ring by comparing with
the Max_replicated_epoch, or if a latest-disk-durable epoch were exposed, it
could be used to determine when a transaction becomes disk-durable.

A new testcase is added to verify the behaviour of the variables.