Bug#18652854 INNODB MEMCACHED ASSERTION 	     `!CONN_DATA->CRSR_TRX' FAILED
PROBLEM
-------

While calling the memcached command 
"flush_all" , we first try to initialize
the connection and transaction .If the 
transaction is in TRX_STATE_NOT_STARTED 
state we do not set CONN_DATA->CRSR_TRX
as NULL.This was causing the assert.

FIX
--- 
If the transaction is in TRX_STATE_NOT_STARTED
state ,we will reuse the transaction. 

[Approved by Jimmy #rb5469]