Bug#18044717: INCORRECT LCP STATE CAUSES AN INTERNAL ERROR
An LCP is initialized by setting the global LCP state to
LCP_INIT_TABLES, then looping over all the ndb tables and
setting the table-specific LCP status to TLS_ACTIVE. If any
table is not ready for LCP, the process continues using a
CONTINUEB signal. This is done to break up the work across
multiple signals to maintain the responsiveness of the
signal processing thread doing the work.

After all the tables are updated, the global LCP status
is set to LCP_STATUS_ACTIVE. Therefore, during the init
procedure, some tables may have a status of TLS_ACTIVE while
the global LCP status is LCP_INIT_TABLES.

While handling node failure of the master during an LCP, the
LCP takeover code checks the global LCP status and, if the
status is LCP_INIT_TABLES, changes it to LCP_STATUS_IDLE. This
results in a case where the global LCP status is LCP_STATUS_IDLE
while some tables may have a state of TLS_ACTIVE.

However, there is an ndbrequire in the node-failure handler
which states that the lcp status cannot be LCP_STATUS_IDLE while
any table has a status of TLS_ACTIVE. This ndbrequire fails and
results in the core.

Removed the ndbrequire to fix this bug. Also added an optimisation
to skip reverting metadata updates while removing a node from a
table if the LCP status is LCP_INIT_TABLES. The revert is not
required in this case since no metadata updates have taken
place yet.