BUG#18466390 INCORRECT BEHAIVIOR WHEN SETTING UP SEMISYNC VARIABLES AS              STARTUP OPTIONS
Semisync didn't work as exepected if below two variables were
set at starting server.
- rpl_semi_sync_master_wait_no_slave
- rpl_semi_sync_master_wait_for_slave_count
The reason was that the variables were changed but the related
internal status were not changed when initializing semisync.

To fix it, internal actions are executed to change the internal status
when initializing semisync at starting server.