The AutoTest run of 'CreateTableWhenDbIsFull' failed due to the DB not being full after all.
This could happen as HugoTransactions::fillTable() gave up
when the first datanode reached 'DbIsFull'. However, there
could still be free DataMemory on other datanodes, such that
the later 'Create Table' succeeded in inserting a row 
on a different (non-full) datanode.

This patch fix ::fillTable() such that it tries more
eagerly to fill DataMemory on all nodes.