WL#7742: BGC: waiting more transactions entering bgc queues
Adds two new options to introduce an artificial delay to make the
binary log group commit procedure wait. This gives a chance that
more transactions are flushed and synced together to disk, thus
reducing the overall time spent to commit a group of transactions
(the bigger the groups the less number of sync operations).
                
These options are named: binlog-group-commit-sync-delay and 
binlog-group-commit-sync-no-delay-count. The former takes as 
input the number of microseconds to wait. The latter, takes as 
input the number of  transactions that the server waits for, 
before deciding to abort the waiting.