Bug #18459944 	NDB REPLICATION : NDB$EPOCH_TRANS DELETE-DELETE-INSERT IS NON-DETERMINISTIC
With NDB$EPOCH_TRANS, delete-delete conflicts were handled like
e.g. update-update conflicts, with the primary rejecting the transaction
and dependents, and realigning the secondary.

This meant that the behaviour w.r.t. following operations on the row(s)
having delete-delete conflicts was dependent on whether they were in
the same or a different epoch.

In the same epoch, they were also considered in-conflict (as the row
in question was 'refreshed' and therefore 'toxic' to following operations).

In different epochs, they were not considered in conflict.

This behaviour was generally regarded as bad for NDB$EPOCH, and so the
same decision is made for NDB$EPOCH_TRANS.

This patch extends ndb_rpl_conflict_epoch_trans to test delete-delete
with NDB$EPOCH and encapsulate the expected result.

One notable enhancement is that randomness is used to have the 
following secondary operations more commonly in a separate epoch.

The transactional conflict handling is modified so that a delete-delete
conflict *alone* is not sufficient to cause a transaction to be
considered in-conflict.