Fix for: 'testNodeRestart -n Bug16772' failure.
In case a NDB_FAILCONF is delayed, or arrives out of order, 
restart of the failed node should await cleanup of nodestatus
on other nodes.

The test expected the failed node to boot in 'no start'
mode, but stall before start phase 3 was reached. However,
this behaviour changed long ago (in 2006) such that the
failed node will now not be restarted at all until 
NDB_FAILCONF has been handled on all nodes.

Test case was adjusted to reflect this new behavior.