Bug#17713871-ASSERT ERROR == DB_SUCCESS, COMMIT_CACHE_REBUILD(), ALTER TABLE, PRBLEM RENAMING
In order to add even more uniqueness to the temporary filename, this patch replaces the LSN in the filename with a more unique another number.  It is just a static global number that is initialized to a random distributed 32-bit number using ut_time() and ut_crc32().  It is then incremented atomically for each temporary file name assigned.  This should not cause performance regression since an atomic_increment is nothing compared to a OS file rename. 

Approved by Marco in RB#5541