Bug #18684389-SPAM ERROR 'FILE NOT FOUND' MSG WHEN 'DATA DIRECTORY' CLAUSE IN 'CREATE TABLE'
The patch for WL7142 caused this regression in which an error message
is written to the log at startup for every remote tablespace that was
created with DATA DIRECTORY.  A 'strict' parameter was added to
Datafile::open_read_only(strict) in order to determine if the message
would be logged.  The function fil_open_single_table_tablespace() tries
to find the file in three possible locations and it always looks at the
default location.  If the file is not there, this strict parameter causes
the spam message.

But by the time the default-location filepath is attempted,
fil_open_single_table_tablespace() has already tried the other two
locations.  So the fix is to print the message only if the file is not
found elsewhere.

Approved by marko in RB#5536