WL#6972: Collect GTIDs to include in the protocol's OK packet WL#6128: Session Tracker: Add GTIDs context to the OK packet
DETAILS
=======

WL#6972: Collect GTIDs to include in the protocol's OK packet
  
This patch collects the GTIDs when they are written to the binary log
and/or when the transaction commits. Whether and at which point the
interception of the GTIDs happen is governed by the variable
SESSION_TRACK_GTIDS, introduced in this worklog. 

If the variable is set to OFF, no GTIDs are collected. If set to
OWN_GTID, then GTIDs generated by the execution of the last statement
submitted, for the given session, are collected. If set to ALL_GTIDS,
then GTID_EXECUTED is collected always, for RO and RW transactions.

WL#6128: Session Tracker: Add GTIDs context to the OK packet

This patch creates an implementation of a session tracker. This
tracker tracks the GTIDs collected through WL#6972 and encodes them in
the OK packet.

This tracker is activated and deactivated using the variable
SESSION_TRACK_GTIDS. If the variable is set to anything else other
than OFF, then the tracker is active and encodes into the OK packet
the gtid set collected during execution of transactions for the given
session.