Bug#18767811-INNODB REPORT "SPACE ID IN FSP HEADER" ERRORS WITH OWN              MULTI-FILE TABLESPACE
In the current mysql-5.5 code, fil_read_first_page() contained an error
in that even though a boolean was sent in to indicate whether this was
the first file node or not, it always read the FSP_FLAGS.  The FSP header
is only valid for page zero, in the first data file of the tablespace.
So fil_read_first_page() in 5.5 is returning junk in the FSP_FLAGS
pointer. This does not matter because the caller that is reading the
system tablespace, the only tablespace with more than one datafile, does
not use the flags returned for datafiles after the first one.

For mysql-5.6, fil_read_first_page() was changed so that in addition to
reading the flags field from the FSP header, it would also read the
space_id using fsp_header_get_space_id(). But this function has a
validity check in it that logs an error message if the space_id is not
correct;

 [ERROR] InnoDB: Space id in fsp header 19742095, but in the page header 0

So this error message is always put into the error log when there is
data in more than one datafile of the system tablespace.  

This error has no effect because the caller ignores the returned space_id
for datafiles after the first one, just like it ignores the FSP_FLAGS.

The fix is very simple.  Do not read these two FSP_HEADER flags if the
datafile is not the first one for the tablespace.

Approved by Marko and Vasil in RB#5630